requirements.txt
# Google API
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread

# Data Processing
pandas
openpyxl

# Fuzzy Matching
rapidfuzz

# Image Processing
Pillow
requests

# Web UI
streamlit
